+++ src/Symfony/Component/Uid/Ulid.php
@@ -157,7 +157,7 @@ class Ulid extends AbstractUid implements TimeBasedUidInterface
             $time = microtime(false);
             $time = substr($time, 11).substr($time, 2, 3);
         } elseif (0 > $time = $time->format('Uv')) {
-            throw new InvalidArgumentException('The timestamp must be positive.');
+            throw new InvalidArgumentException($time, 'The timestamp must be positive.');
         }
 
         if ($time > self::$time || (null !== $mtime && $time !== self::$time)) {